+++ cmd/template/organization/flag.go
@@ -1,0 +1,26 @@
+package organization
+
+import (
+	"github.com/giantswarm/microerror"
+	"github.com/spf13/cobra"
+)
+
+const (
+	flagName = "name"
+)
+
+type flag struct {
+	Name string
+}
+
+func (f *flag) Init(cmd *cobra.Command) {
+	cmd.Flags().StringVar(&f.Name, flagName, "", "Organization name.")
+}
+
+func (f *flag) Validate() error {
+	if len(f.Name) < 1 {
+		return microerror.Maskf(invalidFlagError, "--%s must not be empty", flagName)
+	}
+
+	return nil
+}